Interface + real implementation pattern now the same everywhere
