+++ appBackend/common-service-data/src/jsMain/kotlin/com/gchristov/thecodinglove/commonservicedata/pubsub/PubSubSender.kt
@@ -1,5 +1,6 @@
 package com.gchristov.thecodinglove.commonservicedata.pubsub
 
+import com.gchristov.thecodinglove.kmpcommonkotlin.Buffer
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 
@@ -16,4 +17,18 @@ inline fun <reified T> PubSubSender.sendMessage(
     jsonSerializer: Json
 ) {
     sendMessage(topic = topic, body = jsonSerializer.encodeToString(body))
+}
+
+internal class RealPubSubSender(private val projectId: String) : PubSubSender {
+    override fun sendMessage(
+        topic: String,
+        body: String
+    ) {
+        println(
+            "Sending PubSub message\n" +
+                    "topic: $topic\n" +
+                    "body: $body"
+        )
+        GoogleCloudPubSub(projectId).topic(topic).publish(Buffer.from(body))
+    }
 }
